requirements.txt
aws-cdk-lib
constructs
pandas
scikit-learn
sagemaker
boto3
